Added source redirect function
