+++ Frontend/app/library/index.ts
@@ -1,5 +1,3 @@
 export * from './ickServices';
 
-export * from './createIckForm';
-
 export * from './statServices';
